+++ service-worker.js
@@ -3,7 +3,7 @@
 "use strict";
 
 var CACHE_PREFIX = "monster-ticket-pwa-";
-var CACHE_NAME = CACHE_PREFIX + "741-mobile-20260717-1";
+var CACHE_NAME = CACHE_PREFIX + "744-ticket-reset-20260717-1";
 var OFFLINE_PAGE = "./offline.html";
 var CORE_ASSETS = [
     "./index.html",